Changed nginx templates folder structure
