+++ multienv/web_servers/nginx/templates/template_factory.py
@@ -1,8 +1,8 @@
-from multienv.web_servers.nginx.domain.templates.default_template import \
+from multienv.web_servers.nginx.templates.default_template import \
     DefaultTemplate
-from multienv.web_servers.nginx.domain.templates.laravel_template import \
+from multienv.web_servers.nginx.templates.laravel_template import \
     LaravelTemplate
-from multienv.web_servers.nginx.domain.templates.symfony_template import \
+from multienv.web_servers.nginx.templates.symfony_template import \
     SymfonyTemplate
